requirements.txt
streamlit
nltk
pypdf
faiss-cpu
sentence-transformers
langchain-groq
langchain-community
pandas
openpyxl
python-docx